fix: separated the logical and the visual parts of find specialists page into view and controller
